% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_depth.R
\name{approx_ufg_depth_porder}
\alias{approx_ufg_depth_porder}
\title{Approximates the ufg premises of a list of partial orders}
\usage{
approx_ufg_depth_porder(
  stop_criteria = list(number_iterations = as.integer(1000), number_premises = Inf,
    max_time = Inf),
  porder_observed,
  porder_depth = NULL,
  min_card_ufg = NULL,
  max_card_ufg = NULL,
  input_check = TRUE
)
}
\arguments{
\item{stop_criteria}{(list): list of different stopping criteria. At least
one is not allowed to be infinity. The list must contain a value for
\itemize{
\item number_iterations (eiter Inf or an integer value)
\item number_premises (eiter Inf or an integer value)
\item max time (eiter Inf or an numeric value strict larger than 0.1, the unit is
here minutes)
Note that max_time is not clear yet how it is computed TODO!!!!!!
}}

\item{porder_observed}{(list of square matrix) list of matrices representing
partial orders of the same item dimension, based on these relations the ufg
is computed}

\item{porder_depth}{(list of square matrixes) list of matrices representing
partial orders of the same item dimension, computation of the ufg value of
these partial orders}

\item{min_card_ufg}{(integer or NULL) minimal cardinality of the ufg premise}

\item{max_card_ufg}{(integer or NULL) maximal cardinality of the ufg premise
cardinality}

\item{input_check}{(logical) TRUE when input check sould be done, else set to
FALSE}
}
\value{
returns an approximation of the ufg depth and the total number of
found ufg premises
}
\description{
This function approximates the ufg depth of the partial orders
porder_depth based on porder_observed. Hereby the partial orders are given as
a list of indice matrices. Further one can add a upper and lower cardinality
boundary for the cardinality of the ufg premsies.
}
