% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_general.R
\name{calculate_psi}
\alias{calculate_psi}
\title{Calculates for a subset of objects the minimal intent based on the given
context}
\usage{
calculate_psi(subset_objects, context)
}
\arguments{
\item{subset_objects}{(array): set of attributes}

\item{context}{(matrix): formal context which is used to calculate the intent}
}
\value{
subset (array): the smallest intent (set of attributes) in the FCA
}
\description{
Calculates for a subset of objects the minimal intent based on the given
context
}
