% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_categ_numer_spatial.R
\name{check_boundary_input_cns}
\alias{check_boundary_input_cns}
\title{needed in test_ufg_main_cns
This function checks which point lies in the spatial boundary of the convex
set}
\usage{
check_boundary_input_cns(observed, observed_in_grid)
}
\description{
needed in test_ufg_main_cns
This function checks which point lies in the spatial boundary of the convex
set
}
