% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_function.R
\name{check_input_ccs_1}
\alias{check_input_ccs_1}
\title{Help Function of compute_conceptual_scaling()
Test if all input variables are valid}
\usage{
check_input_ccs_1(
  input_factor = NULL,
  input_ordinal_numeric = NULL,
  input_spatial = NULL,
  input_porder = NULL,
  scaling_methods = NULL
)
}
\arguments{
\item{input_factor}{(NULL or vector of factor values)}

\item{input_ordinal_numeric}{(NULL or vector of ordinal/numeric values)}

\item{input_spatial}{(NULL or list of two dimensional vectors)}

\item{input_porder}{(NULL or list of square same sized matrices)}

\item{scaling_methods}{(NULL or vector containing "porder_edge" or/and
"ordinal")}
}
\value{
logiacl value if input check is passed
}
\description{
Help Function of compute_conceptual_scaling()
Test if all input variables are valid
}
