% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_function.R
\name{check_input_porder_list}
\alias{check_input_porder_list}
\title{Help Function of compute_conceptual_scaling()
Test if all input variables are valid}
\usage{
check_input_porder_list(list_porder)
}
\arguments{
\item{list_porder}{(list of square matrix) list of matrices representing
partial orders of the same item dimension,}
}
\value{
logiacl value if input check is passed
}
\description{
Help Function of compute_conceptual_scaling()
Test if all input variables are valid
}
