% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_function.R
\name{check_input_tipo}
\alias{check_input_tipo}
\title{Input check function to the function test_if_porder
This is just a help function}
\usage{
check_input_tipo(po_candidate, omit_reflexivity)
}
\arguments{
\item{po_candidate}{0-1- Matrix. If entry (i,j)is zero, then i is  smaller
or equal to j. If entry is zero, then this does not hold.}

\item{omit_reflexivity}{(logical) if reflexivity should be omited, so
diagonal entry do not matter}
}
\description{
Input check function to the function test_if_porder
This is just a help function
}
