% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_hierarchical.R
\name{check_mingenerator_hierarchical}
\alias{check_mingenerator_hierarchical}
\title{This is a helpfunction for ufg_1_depth_hierarchical and
ufg_2_depth_hierarchical}
\usage{
check_mingenerator_hierarchical(objset, context)
}
\arguments{
\item{objset}{object set}

\item{context}{(matrix of 0 and 1) representing the hierarchically ordered
data}
}
\value{
logical
}
\description{
This is a helpfunction for ufg_1_depth_hierarchical and
ufg_2_depth_hierarchical
}
