% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_fca.R
\name{compute_context_all_p_orders}
\alias{compute_context_all_p_orders}
\title{Context with all partial orders as intents}
\usage{
compute_context_all_p_orders(n_items, names = (1:n_items))
}
\arguments{
\item{n_items}{is the number of elements of the basic space}

\item{names}{are the names of the n_items elements}
}
\description{
'compute_context_all_p_orders' computes a formal context whose
intents are all partial orders on a set of n_items elements
(PLUS the ALL-relation!): Since every partial order is an intersection
of a set of linear orders (more concretely the set of all linear
extensions), one can can compute the set of all partial orders as the
intents of a formal context where every object is a linear order L and
every attribute is a pair (a,b) and L I (a,b) iff (a,b) in L.
Note that the empty intersection of objects gives the all relation,
and the all relation is not a partial order

#'
}
\examples{
n_items <- 5
steps <- 10000
context_for_n_items_p_orders <- compute_context_all_p_orders(n_items=n_items)
c_orders <- compute_all_partial_orders(n_items=n_items,complemented=TRUE,
list=TRUE)
context <- convert_list_to_context(c_orders,complemented=TRUE)
index <- sample((1:nrow(context)),size=3)
sampled_context <- context[index,]
g <- function(intent,context){0.00001+compute_tukeys_depth(c(intent,1-intent)
, sampled_context)}
tukeys_true_median <- compute_tukeys_median_order(c_orders[index])
tukeys_median_based_on_mc_heuristic <- sample_concept(
context_for_n_items_p_orders,steps=steps,g)
par(mfrow=c(2,1))
plot_relation(tukeys_true_median$median)
dim(tukeys_median_based_on_mc_heuristic) <- c(n_items,n_items)
plot_relation(tukeys_median_based_on_mc_heuristic)

}
