% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_poset.R
\name{compute_context_all_poset}
\alias{compute_context_all_poset}
\title{Context with all partial orders as objects}
\usage{
compute_context_all_poset(n_items, names = (1:n_items))
}
\arguments{
\item{n_items}{is the number of elements of the basic space}

\item{names}{are the names of the n_items elements}
}
\description{
'compute_context_all_poset' computes a formal context whose
objects are all partial orders on a set of n_items elements
(PLUS the ALL-relation!): Since every partial order is an intersection
of a set of linear orders (more concretely the set of all linear
extensions), one can can compute the set of all partial orders as the
intents of a formal context where every object is a linear order L and
every attribute is a pair (a,b) and L I (a,b) iff (a,b) in L.
Note that the empty intersection of objects gives the all relation,
and the all relation is not a partial order
}
\examples{
n_items <- 5
steps <- 10000
context_for_n_items_p_orders <- compute_context_all_poset(
  n_items =
    n_items
)
c_orders <- compute_all_poset(
  n_items = n_items, complemented = TRUE,
  list = TRUE
)

}
