% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implications.R
\name{compute_generator_nominal}
\alias{compute_generator_nominal}
\title{Calculates the generator based on nominal scaled attributes}
\usage{
compute_generator_nominal(subset, list_info)
}
\arguments{
\item{subset}{(vector of (0,1)): 1 represents that the point is within the
subset}

\item{list_info}{(containing data_values): factor attribute of each
observation  (same length as premise)}
}
\value{
the generator of the closure of subset, vectors with elements in 0
and 1, with 1 representing that the observation is in  the set
}
\description{
Based on interordinal scaling this function computes the
generators the conclusion given by subset. The definition of generator is
given by TODO
}
