% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_categ_numer_spatial.R
\name{compute_hull_inner_cns}
\alias{compute_hull_inner_cns}
\title{Computation convex operator for mixed spatial, numeric, categorical data}
\usage{
compute_hull_inner_cns(
  observed,
  observed_in_grid,
  sf_spatial,
  grid_spatial,
  grid_numeric,
  grid_nominal
)
}
\arguments{
\item{observed}{(nx2) the observed values}

\item{observed_in_grid}{(vector) assignment to the grid}

\item{sf_spatial}{(sf-object) converted spatial_grid object to sf-object}

\item{grid_spatial}{(Lx2 matrix) the spatial component of the entire grid
(L is the number of grids)}

\item{grid_numeric}{(vector) the numeric component of the entire grid}

\item{grid_nominal}{(vector) the nominal component of the entire grid}
}
\value{
logical value
}
\description{
This function computes the closure operator for mixed spatial,
numeric, nominal observations
}
\details{
This is based on ufg depth introduced in:
Hannah Blocher, Georg Schollmeyer  (2024+): The union-free generic depth
using formal concept analysis.
}
