% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implications.R
\name{compute_interordinal_conclusion}
\alias{compute_interordinal_conclusion}
\title{Computes the conclusion based on interordinal scaled attribute}
\usage{
compute_interordinal_conclusion(premise, list_info)
}
\arguments{
\item{premise}{(vector of (0,1)): 1 represents that the point is within the
subset}

\item{list_info}{(containing data_values): numeric attribute of each
observation (same length as premise)}
}
\value{
(list): the premise input and the conclusion, two vectors with
elements  in 0 and 1, with 1 representing that the observation is in the set
}
\description{
Based on interordnial scaling this functions computes every
object which lies in the extent set given by the insert objects
}
