% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_loc_sep_statistic}
\alias{compute_loc_sep_statistic}
\title{Location-Separation type test statistic based on Tukeys depth}
\usage{
compute_loc_sep_statistic(corders1, corders2, lambda)
}
\arguments{
\item{corders1}{complemented orders of distribution 1}

\item{corders2}{complemented orders of distribution 2}

\item{lambda}{parameter for setting the threhold c, see above

@return  the value of the test statistic}
}
\description{
'compute_loc_sep_statistic' computes a test statistic for
differences in location or separation of two distributions of order data
based on Tukeys depth: One maximizes the depth of a partial order w.r.t.
distribution 1 under the constraint that the depth w.r.t. distribution 2
is not below a certain threshold c defined by the smallest depth of the
lambda * 100 percent of the observed depth values. For a very small value of
lambda, the test statistic is more or less similar to a generalization of
the M-based#' test, wheres for very high values of lambda one gets
essentially a test that is similar to a generalization of the T-based test
defined in

Jun Li and Regina Y. Liu: New Nonparametric Tests of Multivariate Locations
and Scales Using Data Depth.
Statistical Science , Nov., 2004, Vol. 19, No. 4 (Nov., 2004), pp. 686-696
}
