% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_local_tukeys_depth}
\alias{compute_local_tukeys_depth}
\title{Compute local Tukeys depth}
\usage{
compute_local_tukeys_depth(
  intent,
  context,
  location,
  row_weights = rep(1, nrow(context)),
  col_weights = rep(1, ncol(context))
)
}
\arguments{
\item{intent}{represents the envisaged object given by all its attributes
given as a 0-1 vector. It is also possible to compute depth values for
more objects. In this case the objects should be given as a matrix where
each row corresponds to one
object.}

\item{context}{is a formal context whose objects represent the data cloud
w.r.t. which Tukeys depth is computed}

\item{location}{Object that represents the location w.r.t. which Tukeys
depth should be computed. The location is represented by all its attributes}

\item{row_weights}{it is possible to give every object a weight (with the
interpretation that object with e.g. weight $2$ appear twice as often in the
data set as objects with weight 1)}

\item{col_weights}{it is possible to give every attribute a weight (with the
interpretation that a weighted maximum (minimum) is calculated for
Tukeys outlyingness (depth) such that attributes with higher weights get
more important)}
}
\value{
returns the depth value(s) of the object(s) w.r.t. the data cloud and
w.r.t. the location gibven by the object location.
}
\description{
'compute_local_tukeys_depth' returns the local depth value of
an object represented by its intent (given as a 0-1 vector) w.r.t. a
data cloud represented by a formal context and w.r.t. a given location.
W.r.t. a given location simply means that Tukeys depth is not computed
w.r.t. the whole context but w.r.t. the context that is obtained
if every attribute that the object 'location' does not have, is removed
}
