% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_formal_concepts.R
\name{compute_number_columns_attr}
\alias{compute_number_columns_attr}
\title{Auxiliary function for calculate_conceptual_scaling}
\usage{
compute_number_columns_attr(data_matrix)
}
\arguments{
\item{data_matrix}{(dataframe): each row represents one attribute
(not necessarily two-valued)}
}
\value{
(list): number of attributes needed (all and per attribute),
the column names and the class of the elements
}
\description{
This function is needed to calculate the number of attributes to
represent each column. This depends on the used class of the values in the
column.
}
