% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistica_depth_models.R
\name{compute_probs_depth_model}
\alias{compute_probs_depth_model}
\title{Compute sampling probabilities of a simple statistical depth model
based on a decay function Gamma.}
\usage{
compute_probs_depth_model(depths, scale, p, decay_type = "exp", ...)
}
\arguments{
\item{depths}{a vector with depth-values. The depth-values are assumed to be
between zero and one.}

\item{scale}{is the scale parameter in the model.}

\item{p}{is the power that is applied to the outlyingness-values.}

\item{decay_type}{is the type of the decay function (one of "exp", "inverse"
or "pearson_vii")}

\item{...}{further parameters (e.g., degrees of freedom for the
Pearson type VII model)}
}
\value{
a vector of the same length as the length of dpths which gives the
sampling probabilities according to the model.
}
\description{
'compute_probs_depth_model computes the sampling probabilities
of objects of a formal context based on a decay function Gamma.
}
\examples{
depths <- seq(0.0000000001, 0.99, length.out = 1000)
scale <- 0.05
probs1 <- compute_probs_depth_model(
  depths = depths, scale = scale * 1.5,
  p = 1, decay_type = "exp"
)
probs2 <- compute_probs_depth_model(
  depths = depths, scale = scale, p = 2,
  decay_type = "exp"
)
probs3 <- compute_probs_depth_model(
  depths = depths, scale = scale, p = 0.5,
  decay_type = "inverse"
)
probs4 <- compute_probs_depth_model(
  depths = depths, scale = scale, p = 0.1,
  decay_type = "inverse"
)
probs5 <- compute_probs_depth_model(
  depths = depths, scale = scale, p = 1,
  df = 1, decay_type = "pearson_vii"
)
probs6 <- compute_probs_depth_model(
  depths = depths, scale = scale, p = 1,
  df = 10, decay_type = "pearson_vii"
)
plot(depths, probs1,
  type = "l", lwd = 2, ylim = c(
    min(c(
      probs1, probs2, probs3, probs4,
      probs5, probs6
    )),
    max(c(
      probs1, probs2, probs3, probs4,
      probs5, probs6
    ))
  ),
  main = "Different decay functions", xlab = "depth values",
  ylab = "sampling probabilities"
)
lines(depths, probs2, col = "grey", lwd = 2)
lines(depths, probs3, col = "blue", lwd = 2)
lines(depths, probs4, col = "cyan", lwd = 2)
lines(depths, probs5, col = "red", lwd = 2)
lines(depths, probs6, col = "pink", lwd = 2)
legend(x = 0, y = 0.015, lwd = rep(2, 6), col = c(
  "black", "grey", "blue", "cyan", "red",
  "pink"
), lty = rep(1, 6), legend = c(
  "exp (p=1)", "exp (p=2)", "inverse (p=0.5)",
  "inverse (p=0.1)", "Pearson type VII (df=1)",
  "Pearson type VII (df=10)"
))
}
