% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_general.R
\name{compute_quasiconcave_hull}
\alias{compute_quasiconcave_hull}
\title{Computes the quasiconcave hull of a depth function}
\usage{
compute_quasiconcave_hull(depth_values, context)
}
\arguments{
\item{depth_values}{(vector) depth value of the objects (same order as the
one of the context)}

\item{context}{(matrix with 0 and 1) representing the context corresponding
to the depth}
}
\value{
vector
}
\description{
This is based on ufg depth introduced in:
Hannah Blocher, Georg Schollmeyer  (2024+): The union-free generic depth
using formal concept analysis.
}
