% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_partial_order.R
\name{compute_relation_product}
\alias{compute_relation_product}
\title{Compute the relation product}
\usage{
compute_relation_product(x, y)
}
\arguments{
\item{x}{Represents a relation matrix. Note that has to be a squared matrix.}

\item{y}{Represents a relation matrix. Note that has to be a squared matrix.}
}
\value{
0-1-Matrix: Represents a graph after two steps which are defined by
x and y
}
\description{
Computing the reltion product of two squared, same sized
matrices.
}
