% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_formal_context.R
\name{compute_spatial_scaling_mat}
\alias{compute_spatial_scaling_mat}
\title{Computes the formal context of spatial observation}
\usage{
compute_spatial_scaling_mat(point_matrix, add_column_name = TRUE)
}
\arguments{
\item{point_matrix}{(nx2 matrix):  represents the considered points (objects)}

\item{add_column_name}{(logical): if column names are added}
}
\value{
(list): context (formal context, columns: half_spaces,
row: if point within halfspace)
indexes (describes the index of two points corresponding
to the half-space)
point_matrix (input matrix)
}
\description{
Defining the formal concept of the geometrical formal concept
analysis with object points in R^2 and attributes closed half-spaces and
relation: point is in relation to half-space if and only if point is an
element of this half-space.
Since we want to define a redundant formal-context. It is sufficient to
consider only the half-spaces which bounds goes throw two different points
(objects)
REMARK: This function does NOT work when all point given by point_matrix are
defined upon one line (are collinear)
In particular we assume that all points differ
}
