% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_partial_order.R
\name{compute_transitive_reduction}
\alias{compute_transitive_reduction}
\title{Compute the transitive reduction of a partial order}
\usage{
compute_transitive_reduction(relation_mat)
}
\arguments{
\item{relation_mat}{represents a relation matrix. Note that
has to be a squared matrix.}
}
\value{
The transitive reduction of the relation matrix relation_mat
}
\description{
'compute_transitive_reduction' returns a 0-1-matrix which represents the
order-pairs given reducing all pairs given by the transitivty
}
\examples{
mat <- matrix(0, ncol = 4, nrow = 4)
diag(mat) <- 1
mat[1, c(2,3,4)] <- 1
mat[2, c(3,4)] <- 1
mat[4, 3] <- 1
compute_transitive_reduction(mat)

}
