% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_categ_numer_spatial.R
\name{compute_ufg_grid_cns}
\alias{compute_ufg_grid_cns}
\title{Computes the ufg-depth of mixed cateorical_numeric_spatial data based on an
grid}
\usage{
compute_ufg_grid_cns(
  observed,
  empirical_prob,
  observed_in_grid,
  grid_spatial,
  grid_numeric,
  grid_nominal,
  lower_bound_ufg = 2,
  upper_bound_ufg = 4
)
}
\arguments{
\item{observed}{(nx2) the observed values}

\item{empirical_prob}{(vector) the counting number of duplications for each
observation}

\item{observed_in_grid}{(vector) assignment to the grid}

\item{grid_spatial}{(Lx2 matrix) the spatial component of the entire grid
(L is the number of grids)}

\item{grid_numeric}{(vector) the numeric component of the entire grid}

\item{grid_nominal}{(vector) the nominal component of the entire grid}

\item{lower_bound_ufg}{(integer) value strictly larger than 1
(lower ufg size bound)}

\item{upper_bound_ufg}{(integer) value strictly larger than lower_bound_ufg
(upper ufg size bound)}
}
\value{
returns a list containing the ufg depth, the number of ufg-premises
and a list of possible errors thrown.
}
\description{
Function computes the ufg depht for mixed spatial,nominal and
numeric data based on an grid and an point pattern. For each grid element and
point pattern element the numeric and categorical value must be able to be
evaluated
}
\details{
This is based on ufg depth introduced in:
Hannah Blocher, Georg Schollmeyer  (2024+): The union-free generic depth
using formal concept analysis.
}
