% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_depth.R
\name{compute_ufg_porder_index}
\alias{compute_ufg_porder_index}
\title{Computes the ufg premises of a list of partial orders}
\usage{
compute_ufg_porder_index(porder_list, max_card_ufg = NULL, input_check = TRUE)
}
\arguments{
\item{porder_list}{(list of square matrix) list of matrices representing
partial orders of the same item dimension}

\item{max_card_ufg}{(integer or NULL) maximal cardinalty of the ufg premise
cardinality}

\item{input_check}{(logical) TRUE when input check sould be done, else set to
FALSE}
}
\value{
returns all ufg premises by indicating the one in porder_list
}
\description{
This function computes all union-free generic premises of the
a subset of partial orders. Hereby the partial orders are given as a list of
indice matrices. Further one can add a upper cardinality boundary for the
cardinality of the ufg premsies
}
