% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_weighted_tukeys_depth}
\alias{compute_weighted_tukeys_depth}
\title{Compute the special weighted Tukeys type depth developed in Blocher et al.
2022}
\usage{
compute_weighted_tukeys_depth(
  intent,
  context,
  modus,
  complemented,
  parameters = list(alpha_weight = NULL, beta_weight = NULL)
)
}
\arguments{
\item{intent}{represents the envisaged object given by all its attributes
given as a 0-1 vector. It is also possible to compute depth values for
more objects. In this case the objects should be given as a matrix where
each row corresponds to one
object.}

\item{context}{is a formal context whose objects represent the data cloud
w.r.t. which Tukeys depth is computed.}

\item{modus}{is the location parameter mu for the depth function, cf. Blocher
et al. 2022, equation (2) and the definition of T^mu at p. 25}

\item{complemented}{specifies if the partial orders should be treated
with a complemented conceptual scaling}

\item{parameters}{is a list whith entry alpha_weight and beta_weight which
specify the coefficients of proportionality for the alpha_p,q's and
beta_p,q's of the weighhs in the weighted Tukeys depth, see
Blocher et al. 2022 p. 26}
}
\value{
returns the depth value(s) of the object(s) w.r.t. the data cloud.
}
\description{
'compute_weighted_tukeys_depth' computes the weighted Tukeys
depth for the special case of the set of all partialorders depth introduced
in Blocher et al. 2022
}
\references{
Hanah Blocher, Georg Schollmeyer, Christoph Jansen:
Statistical Models for Partial Orders Based on Data Depth and Formal
Concept Analysis. In: Information Processing and Management of Uncertainty
in Knowledge-Based Systems. Ed. by D. Clucci et al. Springer, pp. 17-30.
}
