% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{convert_context_to_list}
\alias{convert_context_to_list}
\title{Convert a context into a list}
\usage{
convert_context_to_list(
  context,
  complemented = FALSE,
  col_names = NULL,
  row_names = NULL
)
}
\arguments{
\item{context}{The formal context that should be converted}

\item{complemented}{should be true if a complemented context is given (then,
a list of complemented incidence matrices is returned)}

\item{col_names}{Names of the columns of the context}

\item{row_names}{Names of the rows of the context}
}
\description{
'convert_context_to_list' converts a context that represents a partial
order (or an arbitrary homogeneous relation, possibly complemented) into
a list
}
