% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_poset.R
\name{convert_fc_to_list_poset}
\alias{convert_fc_to_list_poset}
\title{Convert a context into a list}
\usage{
convert_fc_to_list_poset(
  context,
  complemented = FALSE,
  col_names = NULL,
  row_names = NULL
)
}
\arguments{
\item{context}{The formal context that should be converted}

\item{complemented}{should be true if a complemented context is given (then,
a list of complemented incidence matrices is returned)}

\item{col_names}{Names of the columns of the context}

\item{row_names}{Names of the rows of the context}
}
\value{
list of posets corresponding to the objects of the context
}
\description{
'convert_fc_to_list_poset' converts a context that represents
a partial order (or an arbitrary homogeneous relation, possibly complemented)
into a list
}
