% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{convert_list_to_context}
\alias{convert_list_to_context}
\title{Convert a list into a context}
\usage{
convert_list_to_context(list, complemented)
}
\arguments{
\item{list}{input list}

\item{complemented}{should the context be returned with complemented
attributes}
}
\description{
'convert_list_to_context' converts a list into a context
}
