% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_general.R
\name{get_weighted_representation}
\alias{get_weighted_representation}
\title{Function to compute the weights of the formal context}
\usage{
get_weighted_representation(x, y = rep(1, dim(x)[1]))
}
\arguments{
\item{x}{at Georg}

\item{y}{at Georg}
}
\value{
list
}
\description{
Function to compute the weights of the formal context
}
