% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{list_to_context}
\alias{list_to_context}
\title{Converts list to context}
\usage{
list_to_context(list, complemented)
}
\arguments{
\item{list}{input list}

\item{complemented}{should the context be returned with complemented
attributes}
}
\description{
'list_to_context' converts a list to a context
}
