% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_general.R
\name{operator_closure_attr_input}
\alias{operator_closure_attr_input}
\title{Defines the closure operator for computing all extents (objects)}
\usage{
operator_closure_attr_input(subset_attribute, context)
}
\arguments{
\item{subset_attribute}{(array): set of attributes}

\item{context}{(matrix): formal context which is used to calculate the
intent}
}
\value{
subset (array): to smallest closure in the FCA based on
subset_object and context
}
\description{
computes the closure operator on the intent set based on a
formal context
}
