% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_general.R
\name{operator_closure_obj_input}
\alias{operator_closure_obj_input}
\title{Defines the closure operator for computing all extents (objects)}
\usage{
operator_closure_obj_input(subset_object, context)
}
\arguments{
\item{subset_object}{(array): set of objects}

\item{context}{(matrix): formal context which is used to calculate the
extent}
}
\value{
subset (array): to smallest closure in the FCA based on
subset_object and context
}
\description{
computes the closure operator on the object set based on a
formal context
}
