% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{plot_order}
\alias{plot_order}
\title{Plot a partial order}
\usage{
plot_order(incidence)
}
\arguments{
\item{incidence}{The incidence relation of the partial order}
}
\value{
The drawing of the partial order in the form of the Hasse diagramm of
the Dedekind-MacNeille completion of the given partial order
}
\description{
'plot_order' plots a partial order by drawing the
Hasse diagramm of its Dedekind–MacNeille completion
}
