% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_poset.R
\name{test_generic_porder}
\alias{test_generic_porder}
\title{Test if the subcontext fc given by partial orders and the scaling method
"edge_non_edge_porder" is a generic premise}
\usage{
test_generic_porder(fc_sub, cardinality_ufg)
}
\arguments{
\item{fc_sub}{(matrix) Formal context describing a subset of partial orders
using the "edge_non_edge_porder" scaling method}

\item{cardinality_ufg}{(integer) number of partial orders, must be same
length as dim(fc_sub)}
}
\value{
logical value stating wether fc_sub is a generic premise
}
\description{
This function test if a premise of partial orders is generic
based on containing
}
