% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{test_if_extendable_to_p_order}
\alias{test_if_extendable_to_p_order}
\title{Check if a binary relation is extendable to a partial order}
\usage{
test_if_extendable_to_p_order(c_order)
}
\arguments{
\item{c_order}{is a binary relation (apposited with its complements)}
}
\value{
TRUE if the binary relation is extendable to a partial order. FALSE
otherwise.
}
\description{
'test_if_extendable_to_p_order' checks if a binary relation can
be extended (in the sense of finding a superset) to a (complemented) partial
order.
}
