% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_partial_order.R
\name{test_if_porder}
\alias{test_if_porder}
\title{Test if matrix represents a partial order}
\usage{
test_if_porder(po_candidate, omit_reflexivity = FALSE)
}
\arguments{
\item{po_candidate}{0-1- Matrix. If entry (i,j)is zero, then i is  smaller or
equal to j. If entry is zero, then this does not hold.}

\item{omit_reflexivity}{(logical) if reflexivity should be omited, so
diagonal entry do not matter}
}
\value{
logical TRUE if po_candidate represents a partial order
}
\description{
Checks if the matrix represents a partial order (thus it
is reflexiv, transitive and anti-symmetric)
}
\examples{
mat_1 <- matrix(0, nrow = 5, ncol = 5)
mat_1[1, 3] <- 1
mat_1[2, 1] <- 1
mat_1[4, 3] <- 1
mat_1[2, 3] <- 1
test_if_porder(mat_1)
test_if_porder(mat_1, omit_reflexivity = TRUE)

}
