% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implications.R
\name{test_interordinal_in_conclusion_newobj}
\alias{test_interordinal_in_conclusion_newobj}
\title{Test if new observation lies in conclusion based on interordinal scaling}
\usage{
test_interordinal_in_conclusion_newobj(subset, obj_numeric_obs, info_list)
}
\arguments{
\item{subset}{(vector of (0,1)): 1 represents that the point is within the
subset}

\item{obj_numeric_obs}{(numeric): observation to test if lies in conclusion}

\item{info_list}{(containing data_values): numeric attribute of each
observation  (same length as premise)}
}
\value{
logical value. TRUE if obj_numeric_obs lies in the conclusion, else
FALSE is returened
}
\description{
Based on interordinal scaling this function tests if a further
object lies in the conclusion of a premise
}
