% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implications.R
\name{test_porder_in_conclusion_newobjs}
\alias{test_porder_in_conclusion_newobjs}
\title{Test if new observation lies in conclusion based on nominal scaling}
\usage{
test_porder_in_conclusion_newobjs(subset, obj_porder_obs, info_list = NULL)
}
\arguments{
\item{subset}{(vector of (0,1)): 1 represents that the point is within the
subset}

\item{obj_porder_obs}{(nominal): observation to test if lies in conclusion}

\item{info_list}{(containing data_values): nominal attribute of each
observation  (same length as premise)}
}
\value{
logical value. TRUE if obj_nominal_obs lies in the conclusion, else
FALSE is returened
}
\description{
Based on nominal scaling this function tests if a further
object lies in the conclusion of a premise
}
