% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_categ_numer_spatial.R
\name{test_ufg_main_cns}
\alias{test_ufg_main_cns}
\title{Testing if and ufg-premise for mixed spatial-numeric-nominal data}
\usage{
test_ufg_main_cns(
  observed,
  observed_in_grid,
  sf_spatial,
  grid_spatial,
  grid_numeric,
  grid_nominal
)
}
\arguments{
\item{observed}{(nx2) the observed values}

\item{observed_in_grid}{(vector) assignment to the grid}

\item{sf_spatial}{(sf-object) converted spatial_grid object to sf-object}

\item{grid_spatial}{(Lx2 matrix) the spatial component of the entire grid
(L is the number of grids)}

\item{grid_numeric}{(vector) the numeric component of the entire grid}

\item{grid_nominal}{(vector) the nominal component of the entire grid}
}
\description{
This function tests if the subset is a ufg-premise for mixed
spatial-numeric-nominal data
}
\details{
This is based on ufg depth introduced in:
Hannah Blocher, Georg Schollmeyer  (2024+): The union-free generic depth
using formal concept analysis.
}
