% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_poset.R
\name{test_ufg_poset}
\alias{test_ufg_poset}
\title{Test if the subcontext fc given by partial orders and the scaling method
"edge_non_edge_porder" is a generic union-free premise}
\usage{
test_ufg_poset(ufg_candidate, input_check = TRUE)
}
\arguments{
\item{ufg_candidate}{(matrix) Formal context describing a subset of partial
orders using the "edge_non_edge_porder" scaling method}

\item{input_check}{(logical, default TRUE) Wether the input should be checked}
}
\value{
logical value stating wether fc_sub is a generic union-free premise
}
\description{
This function test if a premise of partial orders is generic
nion-free based on containing
}
\examples{
relation_1 <- matrix(0, ncol = 4, nrow = 4)
diag(relation_1) <- 1

relation_2 <- relation_1
relation_2[1, c(2, 3, 4)] <- 1
relation_2[2, c(3, 4)] <- 1
relation_2[3, 4] <- 1

list_porder_2 <- list(relation_1, relation_2)

test_ufg_poset(list_porder_2)

}
