% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_hierarchical.R
\name{ufg_1_depth_hierarchical}
\alias{ufg_1_depth_hierarchical}
\title{Computes the ufg premises of cardinality one for hierarchically ordered data}
\usage{
ufg_1_depth_hierarchical(context, weights = rep(1, nrow(context)))
}
\arguments{
\item{context}{(matrix of 1 and 0) representing a formal context}

\item{weights}{(vector) weights of objects}
}
\value{
list of depth value and number of ufg-premises with cardinality 1
}
\description{
This function computes the contributing values for the ufg
depth given by one element ufg premises
}
\details{
This is based on ufg depth introduced in:
Hannah Blocher, Georg Schollmeyer  (2024+): The union-free generic depth
using formal concept analysis.
}
