% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufg_hierarchical.R
\name{ufg_2_depth_hierarchical}
\alias{ufg_2_depth_hierarchical}
\title{Computes the ufg premises of cardinality two for hierarchically ordered data}
\usage{
ufg_2_depth_hierarchical(context, weights = rep(1, nrow(context)))
}
\arguments{
\item{context}{(matrix of 1 and 0) representing a formal context}

\item{weights}{(vector) weights of objects}
}
\value{
list of depth value and number of ufg-premises with cardinality 2
}
\description{
This function computes the contributing values for the ufg
depth given by two element ufg premises
}
\details{
This is based on ufg depth introduced in:
Hannah Blocher, Georg Schollmeyer  (2024+): The union-free generic depth
using formal concept analysis.
It is assuemd that for every category there are at least two different
objects with exactly this category. Additionally it is assumed that on
every level there are at least 3 different categories
}
