% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_betweenness_depth}
\alias{compute_betweenness_depth}
\title{Simple localized depth function based on betweenness w.r.t.a prespecified
mode}
\usage{
compute_betweenness_depth(intent, context, index_modus)
}
\arguments{
\item{intent}{intent of the envisage object}

\item{context}{the underlying context}

\item{index_modus}{the index of the object within the context that represents
the center}
}
\description{
'compute_betweenness_depth' computes a simple depth function
by counting how many points are between the envisaged point and a
center (index_modus). This depth function is not quasiconcave but
it is star-shaped.
}
