% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition_formal_concepts.R
\name{compute_conceptual_scaling}
\alias{compute_conceptual_scaling}
\title{Computes the formal context for ordinal data by the use of dual ordinal
scaling}
\usage{
compute_conceptual_scaling(data_matrix)
}
\arguments{
\item{data_matrix}{(dataframe): each row represents one attribute
(not necessarily two-valued)}
}
\value{
dataframe representing the crosstable/formal context
}
\description{
This function computes the formal context for nominal, inter-
ordinal data points. Here, different data types can be observed.
Information about the scalind methods can be found in:
Ganter, B., Will, R. (2008): Formale Begriffsanalyse,
Mathematische Grundlagen, Springer
}
\examples{
attr_nominal <- as.factor(c("factor_1", "factor_2", "factor_2", "factor_1"))
attr_numeric <- as.factor(c(1.2, 1, 1.6, 2))
compute_conceptual_scaling(data.frame(nominal = attr_nominal,
                      numeric = as.numeric(as.character(attr_numeric))))
}
