% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_categ_numer_spatial.R
\name{compute_hull_inner_cns}
\alias{compute_hull_inner_cns}
\title{This function computes the closure operator for mixed spatial, numeric, nominal
observations}
\usage{
compute_hull_inner_cns(
  observed,
  observed_in_grid,
  sf_spatial,
  grid_spatial,
  grid_numeric,
  grid_nominal
)
}
\description{
This function computes the closure operator for mixed spatial, numeric, nominal
observations
}
