% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_loc_sep_statistic}
\alias{compute_loc_sep_statistic}
\title{Location-Separation type test statistic based on Tukeys depth}
\usage{
compute_loc_sep_statistic(c_orders1, c_orders2, lambda)
}
\arguments{
\item{c_orders1}{complemented orders of distribution 1}

\item{c_orders2}{complemented orders of distribution 2}

\item{lambda}{parameter for setting the threhold c, see above}
}
\value{
the value of the test statistic
}
\description{
'compute_loc_sep_statistic' computes a test statistic for
differences in location or separation of two distributions of order data
based on Tukeys depth: One maximizes the depth of a partial order w.r.t.
distribution 1 under the constraint that the depth w.r.t. distribution 2
is not below a certain threshold c defined by the smallest depth of the
lambda * 100 percent of the observed depth values. Then, by changing the
roles of distribution 1 and distribution 2 and taking a minimum, the
statistic is symmetrized. For a small value of
lambda the test statistic is more or less similar to a generalization of
the T-based test. (This value lambda is that value for which both
optimizations lead to the same maximal value. A value of lambda smaller
than this value seems to be not a reasonable choice). For lambda=1 one gets
essentially a test that is similar to a generalization of the M-based test
defined in Li et al. (2004)
}
\examples{
all_4_c_orders <- compute_all_partial_orders(4, complemented = TRUE,
list = TRUE)
i <- sample((1:219), size = 110)
c_orders1 <- all_4_c_orders[i]
c_orders2 <- all_4_c_orders[-i]
compute_loc_sep_statistic(c_orders1, c_orders2, 0.8)

}
\references{
Jun Li and Regina Y. Liu: New Nonparametric Tests of
Multivariate Locations and Scales Using Data Depth.
Statistical Science , Nov., 2004, Vol. 19, No. 4 (Nov., 2004), pp. 686-696
}
