% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_one_simplicial_depth}
\alias{compute_one_simplicial_depth}
\title{Simple simplicial-type localized depth function based on betweenness w.r.t.
a prespecified mode}
\usage{
compute_one_simplicial_depth(intent, context, index_modus)
}
\arguments{
\item{intent}{intent of the envisage object}

\item{context}{the underlying context}

\item{index_modus}{the index of the object within the context that represents
the center}
}
\description{
'compute_one_simplicial_depth' # computes another simple depth
function that adds for every point of the data clous to allenvisaged points
between this point and the center a one. This depth function is
}
