% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_poset.R
\name{compute_poset_scaling}
\alias{compute_poset_scaling}
\title{testfunktion schreiben und einen kurzen text?}
\usage{
compute_poset_scaling(x, remove_full_columns = FALSE, complemented = FALSE)
}
\arguments{
\item{x}{(matrix)}

\item{remove_full_columns}{(logical)}

\item{complemented}{(logical)}
}
\description{
testfunktion schreiben und einen kurzen text?
}
