% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeys_depth.R
\name{compute_tukeys_depth}
\alias{compute_tukeys_depth}
\title{Computes Tukeys depth of an intent w.r.t. a given formal context}
\usage{
compute_tukeys_depth(
  intent,
  context,
  row_weights = rep(1, nrow(context)),
  col_weights = rep(1, ncol(context))
)
}
\arguments{
\item{intent}{represents the envisaged object given by all its attributes
given as a 0-1 vector. It is also possible to compute depth values for
more objects. In this case the objects should be given as a matrix where
each row corresponds to one
object.}

\item{context}{is a formal context whose objects represent the data cloud
w.r.t. which Tukeys depth is computed.}

\item{row_weights}{it is possible to give every object a weight (with the
interpretation that objects with e.g. weight $2$ appear twice as often in the
data set as objects with weight 1).}

\item{col_weights}{it is possible to give every attribute a weight (with the
interpretation that a weighted maximum (minimum) is calculated for
Tukeys outlyingness (depth) such that attributes with higher weights get
more important).}
}
\value{
returns the depth value(s) of the object(s) w.r.t. the data cloud.
}
\description{
'compute_tukeys_depth' returns the depth value of an object
represented by its intent (given as a 0-1 vector) w.r.t. a data cloud
represented by a formal context.
}
\examples{

context <- fcaR::planets
depth_values <- compute_tukeys_depth(context, context)
table(depth_values)
which(depth_values == max(depth_values))
which(depth_values == min(depth_values))

}
