% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fca_general.R
\name{get_weighted_representation}
\alias{get_weighted_representation}
\title{Computes weighted representation of a data matrix}
\usage{
get_weighted_representation(x, y = rep(1, dim(x)[1]))
}
\arguments{
\item{x}{data matrix}

\item{y}{a further variable. The mean of y in every set of duplicated
x-rows is computed}
}
\value{
list with mean_y: mean of y in the set of the duplicated columns
counts: how often appears the row in the original data matrix X
}
\description{
computes weighthed representation of a data matrix x with
duplicated rows, returns unique(x) together with counts: how often appears
the column, mean_y: mean of y in the set of the duplicated columns
WARNING: the order of the rows generally differs from the order one would get
by applying the function 'unique()'.
}
