% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_fca.R
\name{sample_concept}
\alias{sample_concept}
\title{Sampling a formal concept}
\usage{
sample_concept(context, steps = 1000, f, start_intent = NULL)
}
\arguments{
\item{context}{is the underlying context}

\item{steps}{is the number of steps within the Markov chain Monte Carlo
algorithm}

\item{f}{is the depth function which assigns to every intent a sampling
probability. This functions has to have two arguments: the intent and the
underlying context}

\item{start_intent}{The intent of the formal context at which the Markov
chain starts}
}
\description{
'sample_concept' samples a formal concept of a given formal context
according to a probability function f that assigns to every intent a
corrseponding sampling probability. The algorithm implemented is a Markov
chain Monte Carlo algorithm given in Boley et al. (2010)
}
\examples{
n_items <- 5
steps <- 10000
context_for_n_items_p_orders <- compute_context_all_p_orders(
  n_items = n_items
)
c_orders <- compute_all_partial_orders(
  n_items = n_items,
  complemented = TRUE, list = TRUE
)
context <- convert_list_to_context(c_orders, complemented = TRUE)
set.seed(1234567)
index <- sample((1:nrow(context)), size = 3)
sampled_context <- context[index, ]
g <- function(intent, context) {
  0.00001 + compute_tukeys_depth(c(intent, 1 - intent), sampled_context)
}
tukeys_true_median <- compute_tukeys_median_order(c_orders[index])
depths <- compute_tukeys_depth(sampled_context, sampled_context)

start_intent <- ddandrda:::calculate_psi(
  (sampled_context[which.max(depths), ])[(1:n_items)],
  context_for_n_items_p_orders
)
tukeys_median_based_on_mc_heuristic <- sample_concept(
  context_for_n_items_p_orders,
  steps = steps, g, start_intent = start_intent
)
par(mfrow = c(2, 1))
plot_relation(tukeys_true_median$median)
dim(tukeys_median_based_on_mc_heuristic) <- c(n_items, n_items)
plot_relation(tukeys_median_based_on_mc_heuristic)

}
\references{
Mario Boley, Thomas Gaertner and Henrik Grosskreutz:
Formal Concept Sampling for Counting and Threshold-Free Local Pattern Mining.
Proceedings of the 2010 SIAM International Conference on Data Mining (SDM)
pp.177 - 188
}
